+++ nnetsauce/lazypredict/lazydeepMTS.py
@@ -140,6 +140,15 @@ class LazyDeepMTS(MTS):
             Number of steps ahead to predict (when used, must be > 0 and < X_test.shape[0]).
 
         All the other parameters are the same as MTS's.
+    
+    Attributes:
+
+        models_: dict-object
+            Returns a dictionary with each model pipeline as value
+            with key as name of models.
+        
+        best_model_: object
+            Returns the best model pipeline based on the sort_by metric.
 
     Examples:
 
@@ -153,7 +162,7 @@ class LazyDeepMTS(MTS):
         ignore_warnings=True,
         custom_metric=None,
         predictions=False,
-        sort_by=None,
+        sort_by=None, # leave it as is 
         random_state=42,
         estimators="all",
         preprocess=False,
@@ -187,7 +196,8 @@ class LazyDeepMTS(MTS):
         self.custom_metric = custom_metric
         self.predictions = predictions
         self.sort_by = sort_by
-        self.models = {}
+        self.models_ = {}
+        self.best_model_ = None
         self.random_state = random_state
         self.estimators = estimators
         self.preprocess = preprocess
@@ -217,7 +227,7 @@ class LazyDeepMTS(MTS):
             verbose=verbose,
             show_progress=show_progress,
         )
-        if self.replications is not None:
+        if self.replications is not None or self.type_pi == "gaussian":
             if self.sort_by is None:
                 self.sort_by = "WINKLERSCORE"
         else:
@@ -312,12 +322,12 @@ class LazyDeepMTS(MTS):
             )
 
         # baselines (Classical MTS) ----
-        if X_train.shape[1] > 1:
-            for i, name in enumerate(["VAR", "VECM"]):
+        for i, name in enumerate(["ARIMA", "ETS", "Theta", "VAR", "VECM"]):           
+            try: 
                 start = time.time()
                 regr = ClassicalMTS(model=name)
                 regr.fit(X_train, **kwargs)
-                self.models[name] = regr
+                self.models_[name] = regr
                 if self.h is None:
                     X_pred = regr.predict(h=X_test.shape[0], **kwargs)
                 else:
@@ -335,36 +345,44 @@ class LazyDeepMTS(MTS):
                 else:
                     rmse = mean_errors(
                         actual=X_test,
-                        pred=X_pred.mean,
+                        pred=X_pred,
                         scoring="root_mean_squared_error",
                         per_series=True,
                     )
                     mae = mean_errors(
                         actual=X_test,
-                        pred=X_pred.mean,
+                        pred=X_pred,
                         scoring="mean_absolute_error",
                         per_series=True,
                     )
                     mpl = mean_errors(
                         actual=X_test,
-                        pred=X_pred.mean,
+                        pred=X_pred,
                         scoring="mean_pinball_loss",
                         per_series=True,
                     )
+            except Exception as exception:
+                continue
 
-                names.append(name)
-                RMSE.append(rmse)
-                MAE.append(mae)
-                MPL.append(mpl)
-                if (self.replications is not None) or (self.type_pi == "gaussian"):
+            names.append(name)
+            RMSE.append(rmse)
+            MAE.append(mae)
+            MPL.append(mpl)
+            if (self.replications is not None) or (self.type_pi == "gaussian"):
+                if per_series == False:
                     winklerscore = winkler_score(
                         obj=X_pred, actual=X_test, level=95
                     )
                     coveragecalc = coverage(X_pred, X_test, level=95)
-                    WINKLERSCORE.append(winklerscore)
-                    COVERAGE.append(coveragecalc)
-                TIME.append(time.time() - start)
-
+                else: 
+                    winklerscore = winkler_score(
+                        obj=X_pred, actual=X_test, level=95, per_series=True
+                    )
+                    coveragecalc = coverage(X_pred, X_test, level=95, per_series=True)
+                WINKLERSCORE.append(winklerscore)
+                COVERAGE.append(coveragecalc)
+            TIME.append(time.time() - start)        
+        
         if self.estimators == "all":
             if self.n_layers <= 1:
                 self.regressors = REGRESSORSMTS
@@ -467,7 +485,7 @@ class LazyDeepMTS(MTS):
                     pipe.fit(X_train, **kwargs)
                     # pipe.fit(X_train, xreg=xreg)
 
-                    self.models[name] = pipe
+                    self.models_[name] = pipe
 
                     if self.h is None:
                         X_pred = pipe["regressor"].predict(h=self.h, **kwargs)
@@ -585,8 +603,6 @@ class LazyDeepMTS(MTS):
                             scores_verbose[self.custom_metric.__name__] = (
                                 custom_metric
                             )
-
-                        print(scores_verbose)
                     if self.predictions:
                         predictions[name] = X_pred
                 except Exception as exception:
@@ -653,7 +669,7 @@ class LazyDeepMTS(MTS):
                     pipe.fit(X_train, xreg, **kwargs)
                     # pipe.fit(X_train, xreg=xreg) # DO xreg like in `ahead`
 
-                    self.models[name] = pipe
+                    self.models_[name] = pipe
 
                     if self.preprocess is True:
                         if self.h is None:
@@ -983,24 +999,36 @@ class LazyDeepMTS(MTS):
             }
 
         if self.custom_metric:
-            scores[self.custom_metric.__name__] = CUSTOM_METRIC
+            scores["Custom metric"] = CUSTOM_METRIC
 
         if per_series:
             scores = dict_to_dataframe_series(scores, self.series_names)
         else:
             scores = pd.DataFrame(scores)
-
-        try:
-            scores = scores.sort_values(
-                by=self.sort_by, ascending=True
-            ).set_index("Model")
-        except:
-            pass
+       
+        try: # case per_series, can't be sorted
+            scores = scores.sort_values(by=self.sort_by, ascending=True).set_index("Model")
+            
+            self.best_model_ = self.models_[scores.index[0]]
+        except Exception as e:
+            pass 
 
         if self.predictions:
             predictions_df = pd.DataFrame.from_dict(predictions)
 
         return scores, predictions_df if self.predictions is True else scores
+
+    def get_best_model(self):
+        """
+        This function returns the best model pipeline based on the sort_by metric.
+
+        Returns:
+
+            best_model: object,
+                Returns the best model pipeline based on the sort_by metric.
+
+        """
+        return self.best_model_
 
     def provide_models(self, X_train, X_test):
         """
@@ -1025,13 +1053,135 @@ class LazyDeepMTS(MTS):
 
         """
         if self.h is None:
-            if len(self.models.keys()) == 0:
+            if len(self.models_.keys()) == 0:
                 self.fit(X_train, X_test)
         else:
-            if len(self.models.keys()) == 0:
+            if len(self.models_.keys()) == 0:
                 if isinstance(X_test, pd.DataFrame):
                     self.fit(X_train, X_test.iloc[0: self.h, :])
                 else:
                     self.fit(X_train, X_test[0: self.h, :])
 
-        return self.models
+        return self.models_
+
+class LazyMTS(LazyDeepMTS):
+    """
+    Fitting -- almost -- all the regression algorithms to multivariate time series
+    and returning their scores (no layers).
+
+    Parameters:
+
+        verbose: int, optional (default=0)
+            Any positive number for verbosity.
+
+        ignore_warnings: bool, optional (default=True)
+            When set to True, the warning related to algorigms that are not
+            able to run are ignored.
+
+        custom_metric: function, optional (default=None)
+            When function is provided, models are evaluated based on the custom
+              evaluation metric provided.
+
+        predictions: bool, optional (default=False)
+            When set to True, the predictions of all the models models are returned as dataframe.
+
+        sort_by: string, optional (default='RMSE')
+            Sort models by a metric. Available options are 'RMSE', 'MAE', 'MPL', 'MPE', 'MAPE',
+            'R-Squared', 'Adjusted R-Squared' or a custom metric identified by its name and
+            provided by custom_metric.
+
+        random_state: int, optional (default=42)
+            Reproducibiility seed.
+
+        estimators: list, optional (default='all')
+            list of Estimators (regression algorithms) names or just 'all' (default='all')
+
+        preprocess: bool, preprocessing is done when set to True
+
+        h: int, optional (default=None)
+            Number of steps ahead to predict (when used, must be > 0 and < X_test.shape[0]).
+
+        All the other parameters are the same as MTS's.
+    
+    Attributes:
+
+        models_: dict-object
+            Returns a dictionary with each model pipeline as value
+            with key as name of models.
+        
+        best_model_: object
+            Returns the best model pipeline based on the sort_by metric.
+
+    Examples:
+
+        See https://thierrymoudiki.github.io/blog/2023/10/29/python/quasirandomizednn/MTS-LazyPredict
+
+    """
+
+    def __init__(
+        self,
+        verbose=0,
+        ignore_warnings=True,
+        custom_metric=None,
+        predictions=False,
+        sort_by=None, # leave it as is 
+        random_state=42,
+        estimators="all",
+        preprocess=False,
+        h=None,
+        # MTS attributes
+        obj=None,
+        n_hidden_features=5,
+        activation_name="relu",
+        a=0.01,
+        nodes_sim="sobol",
+        bias=True,
+        dropout=0,
+        direct_link=True,
+        n_clusters=2,
+        cluster_encode=True,
+        type_clust="kmeans",
+        type_scaling=("std", "std", "std"),
+        lags=15,
+        type_pi="scp2-kde",
+        block_size=None,
+        replications=None,
+        kernel=None,
+        agg="mean",
+        seed=123,
+        backend="cpu",
+        show_progress=False,
+    ):
+        super().__init__(
+            verbose=verbose,
+            ignore_warnings=ignore_warnings,
+            custom_metric=custom_metric,
+            predictions=predictions,
+            sort_by=sort_by,
+            random_state=random_state,
+            estimators=estimators,
+            preprocess=preprocess,
+            n_layers=1,
+            h=h,
+            obj=obj,
+            n_hidden_features=n_hidden_features,
+            activation_name=activation_name,
+            a=a,
+            nodes_sim=nodes_sim,
+            bias=bias,
+            dropout=dropout,
+            direct_link=direct_link,
+            n_clusters=n_clusters,
+            cluster_encode=cluster_encode,
+            type_clust=type_clust,
+            type_scaling=type_scaling,
+            lags=lags,
+            type_pi=type_pi,
+            block_size=block_size,
+            replications=replications,
+            kernel=kernel,
+            agg=agg,
+            seed=seed,
+            backend=backend,
+            show_progress=show_progress,
+        )