+++ nnetsauce/votingregressor/__init__.py
@@ -1,0 +1,3 @@
+from .votingregressor import MedianVotingRegressor
+
+__all__ = ["MedianVotingRegressor"]